New command structure
This is the initial commit of the new command structure outlined in
issue #4.  There were major changes to the cli and command base
structures to accomidate the new changes.  Optimist was removed in favor
of mixlib-cli and mixlib-config was added to add some additional
structure around the configuration file and set us up for a bit more
complexity in the structure as we look forward to context setup.

A side effect of adding in mixlib-config was that configuration files are
parsible in either json, yaml, or toml.  I should actually lock that
down to a single type, but at this time I'd like the flexibility to play
around with different formats.  There is a breaking change in that the
config files have been renamed and are checked and loaded in this order:

```
.kafkat.json
.kafkat.yml
~/.kafkat.json
~/.kafkat.yml
/etc/kafkat/config.json
/etc/kafkat/config.yml
```

There were many of the common base interfaces that were moved forward
with minor tweaks, but the addition of mixlib-cli required a new way to
load, list, and interact with the commands.  I've modeled the
interaction after knife (and other ops tooling) which allows you to
list all of the commands and then dive into the individual help sections
to see what the commands actually do and any specific configurations
settings that are specific to the module.

The base command class now defines several global options that each of
the commands inherit.  Among them are configuration overrides that are
merged into the configuration after the file is loaded.

There is some additional refactoring that needs to take place to
decouple utilities that are specific to commands, and place them back
into their appropriate commands, but that will come one by one as we
start preping for the ability to create local commands that can be
loaded in at runtime.