Fix tests, cluster-restart, and update banners
The refactoring broke the tests (which I should have caught earlier) and
parts of the cluster restart command.  Refactored the cluster restart
command to make it work a bit better with the new cli library.  Also
changed the banner to display the full command rather than just the
sub-command.

As part of the testing fixes, I'm now passing ARGV into
the cli object as the rake options were bleeding into ARGV and causing
the tests to fail when the command object - this change also pulled some
of the command initialization logic back into the cli object, which is
where it probably belonged to begin with.

There's probably going to be a bit more cleanup, refactoring, and
testing before I call this one good as I'm not 100% liking how this
looks, however I need to get working on other features.