Airbnb has their own style guide
I just ran across the Airbnb ruby style guide.  I'm not tied to any
specific style and had previously adapted the modified shopify guide to
fill a gap.  I've added the new gem and updated the requirement.  This
does break some of the previous style decisions, so I generated the todo
file to exclude the breaking styles and we can work through them as we
go.  I also added the spec directory back in and am including those
issues for future updates.

I did overide a couple of specific stylistic choices, notably:

```
Layout/AlignParameters:
  EnforcedStyle: with_fixed_indentation
  SupportedStyles:
  - with_first_parameter
  - with_fixed_indentation
  IndentationWidth:

Layout/DotPosition:
  EnforcedStyle: leading
  SupportedStyles:
  - leading
  - trailing
```

IMHO, the alignment of parameters and the dot positions can get messy in
the longer lines, so I tend to prefer this specific indentation style.

Added the tmp directory used by rubocop in the todo generation to the
gitignore file.