Created new run class interface for kafka commands
Added a new interface to the kafka run-class command.  This interface
makes use of the mixlib-shellout library (yes, kindof went crazy with
the mixlib stuff, but they just work).  The RunClass interface is simple
providing a `run!` method which will prepend specific global
configuration values to the command before it is run so the subcommands
will not need to load values such as zookeeper paths and in the future
bootstrap servers.

To test out how the new interface would look with subcommands, I
created the `topic delete` subcommand which takes replicas and
partitions arguments and runs the kafka.admin.TopicCommand with
the appropriate options.

Closes #15