Reduce config load complexity
Some minor cleanup of the config module.  The class cleanup expands the
file paths, and selects the files that exist into a new array.  For
right now we just take the last one in the array that exists (the array
is ordered by precedence).  Though we just take the last right now, this
sets everything up to do a merge of options as we run down the list.
I'm not sure that we need this, but it will be ok to have it in place
just in case.

Added testing around thr config module.