Fix oversight in #39
The recent update of config.rb to reduce complexity in the load! method
introduced a significant error in that it removed the load_file! method
that the cli used to load any custom files that were specified on the
command line.  Instead of adding the method again, I added an optional
paths argument to the load! method.  Unit tests were added around that
specific feature so it won't happen again, but we really need some
integration tests to ensure that the run command works in every senario.

While I was working through that I noticed that there was another
oversight in that we were no longer catching the NotFoundException and
that it was never actually giving a message for the cli to display.
That specific issue has been addressed as well.