Finish style cleanup and fix specs
Closed out the remaining rubocop offenses.  There were a few changes to
the spec files to replace context blocks with describe in cases that
they were referencing class and instance methods.  To clean things up
and also clear some of the multiple top level describes in in the
cluster restart specs, I also split out the cluster restart and session
helpers into there own modules.  This should also help to facilitate the
extension of the sessions helper in the future.

Closes #21
Closes #22
Closes #23
Closes #24
Closes #25